# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATIP public API",
    "version": "1.0.0",
    "description": "Frozen v1 contract. Not financial advice; ATIP is not SEBI registered. Errors: {\"error\": {code, message, request_id, retryable}} (W8)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "ApiKey atk_<id>_<secret>"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "atip_session"
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ],
  "paths": {
...
    "/strategies": {
      "get": {
        "summary": "Strategies of the caller's tenant",
        "operationId": "api_strategies_api_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED"
          },
          "403": {
            "description": "PERMISSION_DENIED / scope missing"
          },
          "429": {
            "description": "RATE_LIMITED (per key, per tenant plan)"
          }
        },
        "x-atip-permission": "strategy:read",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/strategies/{strategy_id}": {
      "get": {
        "summary": "One strategy",
        "operationId": "api_strategy_get_api_strategies__sid__get",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED"
          },
          "403": {
            "description": "PERMISSION_DENIED / scope missing"
          },
          "429": {
            "description": "RATE_LIMITED (per key, per tenant plan)"
          }
        },
        "x-atip-permission": "strategy:read"
      }
    },
...
  },
...
}